refactor: unify agent runtime contracts
